REPO-5405 Allow multiple event type handler
Modify the method getEventType of the interface EventHandler to return
a set of event types.

This way an integrator can create a handler class that implements more
than one of the interfaces provided in the hierarchy simply by providing
an implementation of the method getEventTypes.

The documentation has been updated to reflect this possibility and a sample
implementation of multiple event types handler has been provided in the
event handlers sample application.